added register/auth, comments and updated articles

+++ app/Http/Controllers/ArticleController.php
@@ -4,9 +4,15 @@ namespace App\Http\Controllers;
 
 use App\Models\Article;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class ArticleController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['show']]);
+    }
 
     public function create()
     {
@@ -25,6 +31,7 @@ class ArticleController extends Controller
             'title' => 'required|max:170|min:10',
             'anons' => 'required|min:10',
             'text' => 'required|min:10',
+            'main_image' => 'nullable|image|max:500'
         ]);
 
         $article = new Article();
@@ -32,6 +39,8 @@ class ArticleController extends Controller
         $article->title = $request->input('title');
         $article->anons = $request->input('anons');
         $article->text = $request->input('text');
+        $article->author_id = auth()->user()->id;
+        $article->image = $this->image($request);
 
         $article->save();
 
@@ -54,6 +63,11 @@ class ArticleController extends Controller
     public function edit($id)
     {
         $article = Article::find($id);
+
+        if (auth()->user()->id !== $article->author_id) {
+            return redirect('/')->with('error', 'Это не ваша статья');
+        }
+
         return view('articles.edit')->with('article', $article);
     }
 
@@ -70,6 +84,7 @@ class ArticleController extends Controller
             'title' => 'required|max:170|min:10',
             'anons' => 'required|min:10',
             'text' => 'required|min:10',
+            'main_image' => 'nullable|image|max:500'
         ]);
 
         $article = Article::find($id);
@@ -77,6 +92,15 @@ class ArticleController extends Controller
         $article->title = $request->input('title');
         $article->anons = $request->input('anons');
         $article->text = $request->input('text');
+
+
+        if ($request->hasFile('main_image')) {
+            if ($article->image != 'noimage.jpg') {
+                Storage::delete('public/img/articles/' . $article->image);
+            }
+
+            $article->image = $this->image($request);
+        }
 
         $article->save();
 
@@ -92,7 +116,33 @@ class ArticleController extends Controller
     public function destroy($id)
     {
         $article = Article::find($id);
+
+        if (auth()->user()->id !== $article->author_id) {
+            return redirect('/')->with('error', 'Это не ваша статья');
+        }
+
+        if ($article->image != 'noimage.jpg') {
+            Storage::delete('public/img/articles/' . $article->image);
+        }
+
         $article->delete();
         return redirect('/')->with('success', 'Статья была удалена');
+    }
+
+    public function image(Request $request): string
+    {
+        if ($request->hasFile('main_image')) {
+            $file = $request->file('main_image')->getClientOriginalName();
+            $image_name_without_ext = pathinfo($file, PATHINFO_FILENAME);
+
+            $ext = $request->file('main_image')->getClientOriginalExtension();
+
+            $image_name = $image_name_without_ext . '_' . time() . '.' . $ext;
+            $request->file('main_image')->storeAs('public/img/articles', $image_name);
+        } else {
+            $image_name = 'noimage.jpg';
+        }
+
+        return $image_name;
     }
 }
